Refactored code - Added some CSS on button
